updated according to suggestions given
